requirements.txt
alembic==0.8.10
amqp==2.4.2
apache-airflow==1.10.2
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
billiard==3.5.0.5
bleach==2.1.4
cachetools==3.1.0
celery==4.1.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
colorama==0.4.1
configparser==3.5.3
croniter==0.3.29
cryptography==2.6.1
curlify==2.1.1
defusedxml==0.5.0
dill==0.2.9
docutils==0.14
enum34==1.1.6
facebook-business==3.2.8
facebook-sdk==3.1.0
facebookads==2.11.4
Flask==0.12.4
Flask-Admin==1.5.2
Flask-AppBuilder==1.12.1
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.3.2
flask-swagger==0.2.13
Flask-WTF==0.14.2
flower==0.9.3
funcsigs==1.0.0
future==0.16.0
gitdb2==2.0.5
GitPython==2.1.11
google==2.0.2
google-api-core==1.8.1
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
google-cloud==0.34.0
google-cloud-bigquery==1.10.0
google-cloud-core==0.29.1
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
gunicorn==19.9.0
hmsclient==0.1.1
html5lib==1.0.1
httplib2==0.12.1
idna==2.8
iso8601==0.1.12
itsdangerous==1.1.0
JayDeBeApi==1.1.1
Jinja2==2.10
JPype1==0.6.3
json-merge-patch==0.2
kombu==4.5.0
lockfile==0.12.2
lxml==4.3.3
Mako==1.0.8
Markdown==2.6.11
MarkupSafe==1.1.1
monotonic==1.5
mysqlclient==1.4.2.post1
ndg-httpsclient==0.5.1
numpy==1.16.2
ordereddict==1.1
oauthlib==3.0.1
ordereddict==1.1
pandas==0.24.2
paramiko==2.4.2
pendulum==1.4.4
pip==19.0.3
protobuf==3.7.1
psutil==5.6.1
psycopg2==2.7.7
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
pyasn1-modules==0.2.4
PyHive==0.6.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pysftp==0.2.9
python-daemon==2.1.2
python-dateutil==2.8.0
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==3.0.2
python3-openid==3.1.0
pytz==2018.9
pytzdata==2018.9
PyYAML==5.1
requests==2.21.0
setproctitle==1.1.10
setuptools==40.9.0
six==1.12.0
smmap2==2.0.5
SQLAlchemy==1.2.18
sshtunnel==0.1.4
tabulate==0.8.2
tenacity==4.8.0
text-unidecode==1.2
thrift==0.11.0
tornado==5.1.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.24.1
vine==1.3.0
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.2.1
zope.deprecation==4.4.0
